requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
PyYAML>=6.0
scikit-learn>=1.4
